backend/requirements.txt
# Pipecat framework
pipecat-ai[deepgram,elevenlabs]>=0.0.30

# TLS / certificate bundle
certifi>=2024.2.2

# Web framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
websockets>=12.0

# AI services (Pipecat will use these)
openai>=1.3.0

# Data & Tools
pinecone>=5.0.0
httpx>=0.25.0

# Caching
redis>=5.0.0

# Database (for future use)
sqlalchemy>=2.0.0
asyncpg>=0.29.0

# Utilities
python-dotenv>=1.0.0
loguru>=0.7.0
pydantic>=2.5.0
pydantic-settings>=2.0.0

# Testing & Performance
pytest>=7.4.0
pytest-asyncio>=0.21.0
locust>=2.15.0
